Lazy thread registration
Issue: #11 #12

 - Additional check for `peek_switched`
 - Merged thread_data / thread registry management code
 - TLS-like behavior for POSIX

Leakage and stress test is required.
