+++ modules/thread/get-code.cmake
@@ -5,9 +5,11 @@
 # ---------------------------------------------------------------------------
 cmake_minimum_required(VERSION 3.5)
 
+# message(STATUS ${CMAKE_CXX_PLATFORM_ID})
+
 target_sources(${PROJECT_NAME}
 PRIVATE
     thread/types.h
-    thread/thread_data.cpp
-    thread/thread_registry.cpp
+    thread/compatible.cpp
+    thread/${PLATFORM}.cpp
 )
