+++ modules/win32.vcxproj.filters
@@ -61,14 +61,14 @@
     <ClCompile Include="messaging\queue.cpp">
       <Filter>messaging</Filter>
     </ClCompile>
-    <ClCompile Include="thread\thread_data.cpp">
-      <Filter>thread</Filter>
-    </ClCompile>
-    <ClCompile Include="thread\thread_registry.cpp">
-      <Filter>thread</Filter>
-    </ClCompile>
     <ClCompile Include="messaging\operations.cpp">
       <Filter>messaging</Filter>
+    </ClCompile>
+    <ClCompile Include="thread\win32.cpp">
+      <Filter>thread</Filter>
+    </ClCompile>
+    <ClCompile Include="thread\compatible.cpp">
+      <Filter>thread</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
